Do not add like types to parameters for targeted pessimisation
Summary:
It should not be necessary to add like types to parameters for `<<__SupportDynamicType>>` to be valid. The main situation where this is desirable is when it produces a stronger type of the form `function(~t):u` where `u` is not a like-type, and that this type is essential for some use cases.

For now, let's switch off like types on parameters in targeted pessimisation. To re-enable them, set `--hh-log-level pessimise.params 1` as well as `--hh-log-level pessimise 1`.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D37069279

fbshipit-source-id: 117b6fd0bd68b8e8b7e4b7671c44e50e1e29536d