Fix progress indicator to update typechecked file count only once after remote workers are done
Summary: We began seeing negative progress count during full init. This is because we kept decrementing the completed count well after remote workers were done. This diff adds a check to ensure that we only update the progress once after remote workers return, at which point remaining work should be delegated to the local workers, whose progress update is taken care of elsewhere in the codebase.

Reviewed By: bobrenjc93

Differential Revision: D37087897

fbshipit-source-id: e361a2c97558694e84ead38db44ed6d5cced28e0