type refinements in Typing_core_defs
Summary:
To represent types such as `C with { type T = int }` we
add a list of refinements to the locl type Tclass and a new
decl type Trefinement.

In the initial proposal, refinements can usefully apply
on abstract classes only (concrete classes have their
type members defined). Consequently, it makes sense to
only carry refinements in the case of `Nonexact` class
types. This diff changes `Nonexact` so that it now takes
a list of refinements.

Reviewed By: losvald

Differential Revision: D36805600

fbshipit-source-id: 54546a7d50894af983832a24e94723df83c8fc13