Allow refined dynamic to be passed to SDT function
Summary: When checking the body of a method under SupportDynamicType, a variable may get refined to an intersection type such as `dynamic & nonnull`. The literal treatment of `dynamic` in SDT function application fails to be triggered. A longer term solution would be some subtyping check. For now, let's allow intersections with `dynamic`.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D37137778

fbshipit-source-id: 31d093de680357e7ca2a6742e34c7b807f52b3c4