Refactor daemon mode
Summary: We're only keeping `--daemon` for the few commands that need it, and will soon sink the `--daemon` flag from a global option to subcommand option, for just the subcommands that need it. This diff refactors daemon mode so we dispatch commands centrally first, then enter daemon mode as part of a subcommand implementation if necessary.

Reviewed By: hgoldstein

Differential Revision: D37109099

fbshipit-source-id: 7ef502167f648caff442e4291af2059771514d69