Add "optimistic" mode for extern-worker
Summary:
Add an "optimistic" mode for extern-worker. When storing files or
blobs with extern-worker, you can pass an optimistic flag. If true,
nothing will actually be stored, only the Refs will be generated. You
can perform an exec with these Refs as normal. If the data is already
present in the extern-worker, then the execution will proceed normally
and you've saved a potentially expensive store operation. If the data
isn't present, the execution will fail. In that case, you can then
actually store the data and retry the execution. Different
implementations are not required to support optimistic mode. If they
don't, it behaves as if the flag was false.

The parser uses optimistic mode by default. It can be controlled by
ParserOptimisticStore config option.

Reviewed By: edwinsmith

Differential Revision: D37117903

fbshipit-source-id: 88cf0d30268ab186336df95fdd1f76c4943dc603