Add ability to log positions of `upcast dynamic` in hh server (1/2)
Summary:
Added flag to pass in a function name in ```hh --refactor-check-sound-dynamic``` in order to find all locations where the function was upcasted to sound dynamic. These locations are printed in the hh server. The client continues to print out whether sound dynamic is enabled when running ```hh --refactor-check-sound-dynamic [function name]```.

Since sound dynamic is currently an experimental feature, `--refactor-check-sound-dynamic` is not visible when calling `hh --help`.

Differential Revision: D36884479

fbshipit-source-id: 2ae22307958149759582692a77ba9703bf442e9f