Newline-aware Lexer
Summary: The next step in building the assembler is being able to parse opcodes. In HHAS, an instruction in a function body occupies a whole line. So in order to be able to parse function body instructions, we need a lexer that is aware of newlines and can spawn mini-lexers based on these newlines. With this construct we can parse and assemble function bodies, by calling `advance_past_newline` while our peek-ahead is not a `}`.

Reviewed By: hgoldstein

Differential Revision: D37124182

fbshipit-source-id: a246e3dec9dec4157d1f7ba45e9dd104a453286a