Fix coeffect check local off by one indexing in prologues
Summary:
Prologue logic is given a number of passed arguments via `argc`, which includes
the unpack argument. This unpack argument is popped off the stack if the callee
does not accept variadics. The coeffect logic did not account for this and if
unpack args were given, its indexing into locals would be off by one.

Instead of adjusting the coeffects logic, fix this bug by updating the `argc`
value. After the recent calling convention changes the original `argc` value is
no longer needed.

This also simplifies logic for the next diff that moves argument type checks to
prologues.

Reviewed By: ricklavoie

Differential Revision: D37160577

fbshipit-source-id: 711a866790943e3cf38d828a2afcc28b5319a615