Fix HHBBC monotonicity violation for FCallClsMethods.
Summary:
The monotonicity requirement basically requires that refining
an input type should never result in a worse output type. For the
FCallClsMethods (and FCallClsMethodM in particular), we were violating
that. If the input type could be an object (or others), we analyze the
call as normal and push some return type. However, if the input type
isn't valid, we just call fcallUnknownImpl (which does the pessimistic
thing and push InitCell). So, if the input is originally ?Obj, we'll
process it normally. When we refine the input to InitNull, we now push
InitCell, so monotonicity is violated. If we know the input isn't
valid, we must properly treat it as unreachable.

When I experimented with doing limited bottom-up analysis, this caused
monotonicity assertions right away.

Reviewed By: paulbiss

Differential Revision: D37193257

fbshipit-source-id: df73e296ec9f6a2ab8d21af93bf57817ba99655e