Initial infrastructure for fine-grained dependency tracking
Summary:
This diff adds some initial infrastructure for a new mechanism for collecting dependencies in `Typing_fine_deps`, more fine-grained as the existing implementation in `Typing_defs`.

It also adds the field `dchild` to the type `Decl_env.env`. It complements the field `droot` of the same record, by indicating which member/child of `droot` is actually being type-checked.

As this new mechanism may only be used for the Sound Dynamic project, it is likely that it may be removed at some point in the future. It also seems that the dependencies collected using this new mechanism will not be a strict superset of those collected by `Typing_defs`. To this end, the new mechanism is implemented in parallel to `Typing_defs`, rather than by modifying/enhancing the latter.

Reviewed By: CatherineGasnier

Differential Revision: D36970787

fbshipit-source-id: 886b7b60639819e70a492f80870d3511121ebb0c