Add option for recording fine-grained dependencies
Summary:
This diff adds a new option `tco_record_fine_grained_dependencies` to the type `GlobalOptions.t`. It is used in subsequent diffs such that when enabled, a new mechanism for collection and recording dependency information is used, which is more fine-graine than the existing one.

Alternatively, it would be possible to add another "mode" to the type `Typing_deps_mode.t`, but this would require more changes in `Typing_deps` itself. I've therefore decided against this and went for a standalone option instead.

Reviewed By: CatherineGasnier

Differential Revision: D36972062

fbshipit-source-id: 38b5cfeb27d44db3246486f7968daa82afcd1e7a