Inferred lambda types should have supportdyn in SDT mode
Summary:
We're moving to a model where we use `supportdyn<t>` to indicate that a function type `t` supports calling with dynamic. Unfortunately this isn't reflected in the function type inferred for a lambda. This diff fixes that.

Subsequent diffs will remove the legacy `ft_flags_support_dynamic_type` bit on function types, in favour of using `supportdyn<_>` only.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D37188177

fbshipit-source-id: 7e50d07a47679d7fb4214b40f10f75da2096c5a6