Pass dependencies to Typing_fine_deps functions
Summary:
This diff ensures that when `GlobalOptions.tco_record_fine_grained_dependencies` is enabled, appropriate functions from `Typing_fine_defs` are invoked on all dependencies.

Here "all dependencies" means that the set of dependencies collected this way is a a strict superset of those collected via `Typing_defs.add_idep`: for the time being, we collect the same dependencies as the existing mechanism for the sake of debuggability. This will change in the future.
Additional dependencies (as compared to the original dependency collection mechanism) arise from the fact that we already support methods as dependents a this points (i.e., we can track the dependencies of individual methods, which is not the case in the graph created by `Typing_deps`). Therefore, this diff simply adds calls to `Typing_fine_deps` functions wherever alls to `Typing_deps.add_idep` currently happen.

The changes made by this diff are intended to be easily removable, which is why I eschewed for example creating variables for information needed by both the call to `Typing_deps.add_idep` and the newly added one.

Reviewed By: CatherineGasnier

Differential Revision: D36973314

fbshipit-source-id: acb0ffa08a50cb8079e52c6b7db20dfd5d44ac5f