Don't branch on m_nextKI in HashCollection::shrink
Summary:
HashCollection::shrink has an if clause that checks if the m_nextKI on the underlying dict is 0. The intent of this early returning block is to replace the underlying dict with the static empty one when m_size and m_nextKI are both 0.

This code is very cold, and in fact very hard to reach (see new unit test). We can remove this unnecessary micro-optimisation and fall through to code that will still shrink the collection, just not down to a capacity of 0.

With this change, all code touching m_nextKI doing is one of:
1) initializing it to 0
2) copying it from one container to another during e.g. grow/shrink operations
3) updating m_nextKI after an insert or erase at an int key
4) setting it to m_size after a full renumbering of the dict keys
5) is the RO::EvalDictDArrayAppendNotices code in VanillaDict::AppendMove

This means that once the RO::EvalDictDArrayAppendNotices notices go to 0 we will be able to safely delete m_nextKI (replacing it with the computed value in the only real user of it (case 5 above) that currently logs differences with the computed value it's comparing to).

Reviewed By: paulbiss

Differential Revision: D37117564

fbshipit-source-id: cfa59cc44aaf437c6718d4e19f817848fc7e93c8