Persisting of dependencies using SQLite
Summary:
This diff adds submodules `Backend` and `SQL_persistence` to `Typing_fine_deps`, which implement the actual logic behind  `Typing_deps_add_fine_dep`. The former module implements a cache of dependencies that is flushed to disk using the latter module.

One file is used per worker, and caches are flushed to disk during a finalization operation when a worker is done type-checking. The individual per-worker files need to be merged and deduplicated, which is implemented in subsequent diffs.

Reviewed By: CatherineGasnier

Differential Revision: D36973629

fbshipit-source-id: 45c260c986a7bb25037c9617b705b5a97b46daba