Treat enums as unenforced
Summary: Enums are enforced only at the base type. For the purposes of Sound Dynamic, we should therefore treat them as unenforced. There are ways of making use of this partial enforcement via intersection types, which we should explore. Depending on the outcome of our pessimisation experiments, we might want to revisit this choice and treat enums as enforced but live with the mild "lie" that base-type enforcement induces. But for now, let's work with the assumption that enums are unenforced.

Differential Revision: D37237548

fbshipit-source-id: 49729b8dd40d0b5a5e3f031ca1332aace69804e1