Propagate module membership declaration (SetModule) to NAST
Summary:
This diff is just some of the changes from the next diff D36647765 separated out on its own.

The logic that makes hover and go to def possible in the IDE uses a
TAST visitor, so module membership declaration needs to be present in the NAST.
This enables the behavior where, when a module membership declaration is
clicked, the module definition is brought up. I am not sure why it doesn't need
to be in the TAST itself, only the NAST, but maybe it has to do with visitor
inheritance?

Reviewed By: jamesjwu

Differential Revision: D37200769

fbshipit-source-id: 417d96331fd1abdcb23866a52c308507893f43f1