Use a runtime option to optionally throw if DB doesn't already exist
Summary:
By default, if a Facts DB doesn't already exist, native Facts will attempt to load the repo from scratch and create the DB.

In some situations, a running HHVM webserver will get pinged and will try to respond to requests, even though a Facts DB doesn't exist yet.

1. Perhaps we should crawl the repo and build a DB on demand so we can autoload and serve these requests, or
2. Perhaps we should fail the request because we're still waiting to download the DB from saved state, and we have no business running code out of a repo that's not ready yet.

Previously, we always attempted to crawl the repo and build a DB on demand, in an attempt to make the request succeed at all costs. But a repo may be too large for this to be feasible, and any request that tries to do this will inevitably time out anyway. This diff adds option (2), alongside a runtime option to throw if a DB doesn't already exist.

Reviewed By: ricklavoie

Differential Revision: D37085646

fbshipit-source-id: 4b08d364441ed5f41cfbc8ed26cdc75eebefccfb