reimplement go-to-def on `<<__Override>>` for folded decls
Summary:
The IDE feature "go-to-def" works on `<<__Override>>` and jumps to the method being overridden. It's currently implemented in `ServerSymbolDefinition.go` but unfortunately relies on the order of the ancestors provided by `Typing_classes_heap.Api.all_ancestor_names`.
That order is different for shallow decls and folded decls. For shallow decls, it's the order of the linearization, which works well to find the overridden method, while for folded decls it's the alphabetical order.

Therefore, this diff reimplements this feature, by providing a new helper `Decl_provider.get_overridden_method`. The core of that method is written in Decl_inherit and reuses the parent traversal logic provided by `Decl_inherit.OrderedParents`.

Reviewed By: andrewjkennedy

Differential Revision: D37066077

fbshipit-source-id: 18db36c829babe89910592d3fdb68fa30b17663d