Make gdb think stderr is going to a terminal, so tui can be enabled
Summary:
When you run `hhvm_wrapper.php --gdb` to try and run a Hack file in GDB, GDB does not recognize the stderr as a tty, so you cannot enable TUI if desired.

The current main workaround is to just run `hhvm_wrapper.php --print-command` and use the result as arguments to your own invocation of `gdb --args`. However, we can save a step or two by fixing the wrapper's invocation itself.

Reviewed By: mofarrell

Differential Revision: D37259121

fbshipit-source-id: 821a7a77b938884410bdb38d701782f90e8bd9ea