Move logic logic for typing classes to Typing_class
Summary:
`Typing_toplevel` contains the function `class_def` and almost 2000 LOC specific to the implementation of that function. This diff moves `class_def` and its helpers into a new file called `Typing_class`.

Beyond the desire to reduce the number of overly large files, this change creates a reasonable module to expose functions that check individual members of a class, rather than the whole class (Members are not toplevel definitions, and adding function for checking them individually to the interface of `Typing_toplevel` is contradictory, whereas adding them to `Typing_class` is much more sensible.)

The only changes that go beyond just moving code around are the following:

1. The two functions `merge_hint_with_decl_hint` and `merge_decl_header_with_hints` are needed by functions in `Typing_class` and `Typing_toplevel.fun_def`. I moved them to `Typing_helpers` to make them available in both modules.
2. Similarly, the function `is_literal_expr` is needed by modules. Its old comment suggests that "This should agree with the set of expressions whose type can be inferred in Decl_utils.infer_const". I've now enforced this by calling the latter function directly and leaving a helper stub called `is_literal_with_trivially_inferable_type` in `Typing_class` and `Typing_toplevel`.

Reviewed By: andrewjkennedy

Differential Revision: D37278990

fbshipit-source-id: 6a0d5e301f0006ac5e250ab079b21f100d310a6e