Improve inference around like-type call rule
Summary:
D33189708 (https://github.com/facebook/hhvm/commit/101ebb308940e2f7005847e62fc7698ce02e642c) and D33915660 (https://github.com/facebook/hhvm/commit/660d38efb66947ee5daafd5b2fff30df7e137c18) implemented a special rule for calling `supportdyn<function (t1, ..., tn) : t>` functions, instead of using an actual intersection type. If type checking with the static type fails, the rule checks that the arguments are dynamic-aware subtypes of dynamic. This is all that is needed for dynamic soundness, but to ensure that important type errors are kept (e.g., calling `~int` with `~string`), it removed the like type from the argument. This was done with a simple syntactic check that did not work when the type of the argument had a complicated structure, or was a variable with lower bounds.

Here, instead of removing the ~ from the arguments, we add it to the parameter, which has the same results. When calling a parameter with type `t` with an argument of type `~u`, and `u <:D dynamic`, then checking `~u <: ~t` will reduce to checking that `dynamic <: t|dynamic and u <: t|dynamic`. The first is trivial, and the second should keep the same errors as `u <: t` since we are not using dynamic-aware subtyping.

Reviewed By: andrewjkennedy

Differential Revision: D37311931

fbshipit-source-id: dafa1142b7b8b1afe794f6c3393e1c976b402af4