Allow public at toplevel declarations
Summary:
This diff allows public as a visibility on toplevel definitions when modules are enabled.

We start by allowing it as an optional keyword, which no one will care about initially. We'll then enforce in the typechecker that all toplevel code in modules must have a visbility (public or internal) attached to it.

Reviewed By: oulgen

Differential Revision: D36459271

fbshipit-source-id: 274c40f296b4bf38cc4c3910116d6976b722fc89