Do not perform cfg opts on Enter opcode
Summary:
Do not attempt to optimize Enter opcode in hhbbc. Enforce that it always points
to the main func entry.

Ideally the Enter opcode would not have an offset immediate at all, but that
requires more work, as many places in JIT treat Enter as Jmp. That cleanup will
be done later. At this moment we just need the offset to be always zero.

Reviewed By: ricklavoie

Differential Revision: D37326674

fbshipit-source-id: 214dbd0dd21107e6c58279621c9aa961838c0d78