Fix target pessimisation test to use batch files
Summary:
It's important that we use the `--batch-files` option when running `hh_single_type_check` on a directory, otherwise the files are checked as a group, and we end up with `name already bound` errors.

So scrape the `.out` file for pessimisation log entries. For now, we don't delete the file in order to debug pessimisation problems more easily.

Reviewed By: vassilmladenov

Differential Revision: D37318174

fbshipit-source-id: 3fcd8b639e23b6488ece78c08277f47e32a5ba63