Expose `HH\is_class_meth` to `ext_array`
Summary:
`HH\is_class_meth` is a pseudofunction: there's no PHP definition for it as it's compiled to a bytecode operation. Let's move it to `pseudofunctions.hhi`, which is included when typechecking systemlib.

While I was here I cleaned up some of the documentation for functions already in `pseudofunctions.hhi`

Reviewed By: aorenste

Differential Revision: D37251330

fbshipit-source-id: 144b6a9bf248b3aa715227e5a4615dab7a884f24