Ban aliases in catch blocks
Summary:
An alias is completely ignored by HHVM in catch blocks, leading to users thinking that their code will catch exceptions that it ignores.

This might be something we can support in the future when decls-in-compilation enables aliases in more positions, but today it's safer to update hh to reflect the runtime.

Differential Revision: D36915826

fbshipit-source-id: e60d5ed19314878f57fed8479faf11f31f5f520d