Remove use of BumpSliceMut from Instruct
Summary:
Using BumpSliceMut in Instruct meant that we couldn't use PartialEq to compare two Instructs that had different lifetimes (mumble mumble rust lifetimes "lifetime b flowing into lifetime a").

This could have been solved by having a specialized PartialEq defined for Instruct (and Opcode) that explicitly understood comparing separate lifetimes - but either that's dangerous because we need to write a custom PartialEq or it's annoying because we need to write/support a custom derive macro to generate it for us.

But BumpSliceMut is barely used in Instruct - so instead just stop using it.  Really the only place it was needed was in the label_rewriter - so that was pretty straightforward to have it just rewrite the Instruct instead.

Differential Revision: D37335558

fbshipit-source-id: c2d0d0b481be98edadd99e6e21de23d098eba646