Remove the global literal array and string tables (take 2)
Summary:
If we wish to build a repo-file in a more distributed way, it becomes
difficult to support single global literal array and string tables. We
already have support for per-unit literal and string tables (this is
needed for Sandboxes), so just always use those. This change actually
tends to produce more compact bytecode, since the literal ids are, on
average, much smaller now.

We did not support lazy loading of strings and arrays for the per-unit
tables, so add support for that. We do no longer need the "use global
ids" flag for BlobDecoder or UnitEmitter, so that can be removed.

This is a retry of an earlier diff, which had to be reverted because
of some strange warmup issues. Two changes have been made to attempt
to "fix" this. First, encoding LowStringPtr now uses the unit's string
table instead of always encoding the string directly. This greatly
shrinks the size impact of this change. In fact, we can go one further
and use the unit's string table for *all* StringDatas (which helps a
bit more). The second change is an optimization for lazy loading. When
deserializing the unit-emitter and creating the string table, check if
the string corresponds to an already existing static string. If it
does, set that static string in the table directly (no lazy
loading). If not, store the offset as usual (the normal lazy
case). This means that common strings, which are almost certainly
going to have been created already, won't need to be loaded later
(which is more expensive).

Using the unit's string table during encoding causes some ordering
dependencies, so add infrastructure to BlobEncoder/BlobDecoder to
handle cases like this.

Reviewed By: edwinsmith

Differential Revision: D37303970

fbshipit-source-id: 1531414132daba963ba1cccbcfc39581d182088b