Remove global RAT array table (take 2)
Summary:
Similar to the literal string and literal array tables, remove the
global RAT array table and replace it with per-unit tables. While I'm
at it, add support for lazy loading of the RAT arrays as well.

There's no need for the RepoAuthType logic to be split into 3
different sets of files/headers, so consolidate it all into one.

This is a repeat of the first attempt at this change. The only
difference is the addition of tl_unit, which is used in a similar
manner of tl_unitEmitter for lazy loading.

Reviewed By: jano

Differential Revision: D37328235

fbshipit-source-id: d85b6a54433362c78c63a2d1aef5f1a85bca3cb6