Fix named locals in disas
Summary:
Fix NLA printing logic in disas, which was not handling a situation where hhbbc
optimized away the local name.

Fix MK - MEL/MPL printing logic that was using a wrong field from the union,
which happened to be the same as the named local, which could be optimized away
by hhbbc. Use the correct field -- the effective local id. Note that this
still doesn't round trip as hhas doesn't yet support full representation of
named locals here. But at least it produces legit executable code and doesn't
segfault.

Reviewed By: ricklavoie

Differential Revision: D37386103

fbshipit-source-id: 7f047a69c2e9a366c044c041beb543eacc4aa890