get bespoke type structure fields by offset calculation
Summary:
Added logic to access fields from bespoke type structures using field offsets instead of always the generic dispatch method.

New IR instruction: `LdTypeStructureValCns`

This instruction is an optimization on `LdTypeStructureVal`. When a `key` is a const string value, then we know the exact field that is being accessed. Since bespoke type structures have a fixed layout, we can calculate the offset to that field and directly load the value stored at that offset.

`LdTypeStructureValCns` loads the memory at the offset to the `dst` register and does not check whether the field is valid on the type structure. (e.g. the `alias` field always exists but can be `nullptr`). So after accessing the field, in `simplify.cpp` there is additional logic to check that the field is a valid type. We branch if the field is invalid.

Other changes:
* modified `CheckNonNull` to take in more input types
* created new type `DTypeStructElem` to handle specific output return types for type structure fields (Vec and Dict are separated since `CheckNonNull` requires a single type)

Reviewed By: paulbiss

Differential Revision: D37134109

fbshipit-source-id: b5bbac082e73f6790efcece052d669e1cd0e9dd5