Improve subtyping completeness
Summary:
When making an IsSubtype proposition during dynamic-aware subtyping with `dynamic` as the supertype, the `dynamic` needs to become `supportdyn<mixed>`. If this doesn't happen, subsequent, non-dynamic-aware calls do not know when to resume dynamic-aware subtyping, if the `dynamic` ends up as an upper bound on a type variable. Previously, the code just looked for a literal dynamic, now it searches through unions, which is still potentially brittle, but less so, and handles the following kind of example.
If we have
`#0 <: supportdyn<mixed>`
and then do the following,
`#0 <D: dynamic | nonnull`
we end up with `dynamic | nonnull` as an upper bound on `#0` if we fail to detect `dynamic` when making the `IsSuptype`, since when we add the bound later we don't use dynamic-aware subtyping, and we don't have `supportdyn<mixed> <: dynamic` under non-dynamic-aware subtyping. By instead placing `supportdyn<mixed> | nonnull` into the subtype, the bound can correctly be identified as redundant later on.

Reviewed By: andrewjkennedy

Differential Revision: D37418257

fbshipit-source-id: 813920b03846eb934522d127f5a8ef8399aed2db