Add a bytecode instruction that enables verification of implicit context in memoized wrappers
Summary:
In order to build dynamically enforced implicit context, we will need to check certain properties about the implicit context in memoized wrappers. This verification, based on the state of the implicit context and the calling context, will raise warnings or throw exceptions. Once such exception is included in the diff.

This diff adds a generalized bytecode instruction as a starting points.

Reviewed By: ricklavoie

Differential Revision: D37011085

fbshipit-source-id: 349958a8bbd48f3d48253565175b8026f3e68d08