Deal with interfaces properly in HHBBC's type-system (take two)
Summary:
Interfaces make HHBBC's type-system be non-monotonic. By monotonicity,
I mean that if a <= A, and b <= B, then a|b <= A|B always (and the
same applies for intersection). This isn't the case if you're unioning
or intersecting together interfaces. The union or intersection of an
interface with a different class may not be representable by a single
class. Normally we'd deal with this by finding a super type which is a
single class (or failing that, using TObj). This is what breaks
monotonicity since that super type may not implement the
interface. So, by widening the type, we've lost subtype information.

This problem has been long known, and we've hacked around it by
various means. The union and intersection operators are pessimistic
when encountering an interface, and we use loosen_interfaces (which
just drops the class info if an interface) liberally. Despite this, we
still occasionally get monotonicity violations cropping up (usually
with refinement errors in the index), and we pessimize yet more.

Fix this all for good by properly modeling interfaces in HHBBC's
type-system. As I said above, no type-system which can only ever
represent a single class will work. We expand the type-system to be
able to represent the intersection of classes explicitly. The
representation is the list of all the classes (with redundancies
removed) that the class is a subtype of.

For example, if A is a class, and I is an interface, then A&I means
the class is both a subclass of A and implements I. We can now
represent the union or intersection of an interface without having to
widen the type "too far". The representation preserves the list of
classes it is a subtype of, preserving monotonicity.

Constructing this representation is relatively easy. In general, given
a group of classes, find the base class they have in common (if any),
and the list of interfaces they all implement. This list of common
classes is the representation (given a suitable canonicalization). For
unions, the group of classes is the union of both sides. For
intersection, the group of classes is the intersection on both
sides. The procedure is the same otherwise.

The representation has the benefit of being compact (avoiding the blow
up of an explicit list of subclasses, for example). It also can only
shrink under the common operation of unioning. Again, this avoids
pathological blow up of unioning large sets.

To preserve unique type representations, we impose a canonical form on
the above list. All classes on the list must be "couldBe" each other
(that is, their intersection is non empty). If not, the list is
actually Bottom, because no class could be a subtype of
all. Furthermore, no class on the list can be a subtype of another. If
a class is a subtype of another, the larger class is redundant, since
being a subtype of the smaller class implies being a subtype of the
larger class. Finally, the classes on the list are sorted. The sorting
is arbitrary but must be consistent. We order the classes by their
size, so that "smaller" classes come first. Note: if the union or
intersection is a subtype of a single class or interface, this
representation becomes the existing single class representation (the
list will only have one element).

DCls already can represent an exact class, or a sub class. Add a third
case, the above intersection of classes. The Tag of DCls was becoming
cumbersome, so replace it with explicit query functions.

To implement subtypeOf() and couldBe() correctly on interfaces, we
need some pre-calculated information about the relationship between
interfaces. This information is calculated lazily but is stored on
ClassInfo.

Note: The meaning of exact/sub for an interface is unintuitive. An
"exact" interface is the interface class itself, which has no parents
and implements nothing. A "sub" interface are the implementors of that
interface (which does not include the interface class itself). As a
result, a Obj=I is not a subtype of Obj<=I (in fact, they're not
even "couldBe"). The type system was not, in general, getting this
correct, so fix all of that (and try to explain the distinction).

To further ensure type uniqueness, canonicalize interfaces when we
create a Type with them. Two interfaces are equivalent if they have
the same set of implementators. Among a set of equivalent interfaces,
we canonicalize to the one with the lesser name. If an interface has
no implementations at all, it is equivalent to TBottom.

The type-system also wasn't getting unresolved classes quite right. An
unresolved class can never be a resolved class (it would be resolved
if it was).

We need to be able to use resolve_method on a DCls with the new
intersection representation. This necessitates adding a new variant to
res::Func, representing an intersection of func families. This behaves
like a normal func family, except there's multiple and you must
"intersect" the results of all of them.

With all of this done, we can now delete all of the loosen_interface()
calls and what not. In fact, there needs to be no special handling for
interfaces outside of the type-system at all now.

Reviewed By: jano

Differential Revision: D37433315

fbshipit-source-id: 49cb6709e8faa146c619fc1cf1a9ce64b1281adb