Add support for exponential backoff retries to extern-worker
Summary:
Some backend implementations for extern-worker can throw an error if
they are sufficiently overloaded. Right now this is treated like any
other error and generally results in us using the fallback
implementation. However, unlike other errors, this is generally
transient and retrying may succeed.

In this situation, have the backend throw a new exception (called
Throttle). This indicates the backend is overloaded and the caller
should wait a bit and retry. The main Client implementation
automatically catches this, sleeps for some amount of time, then
retries. The number of times to retry (if any) is given by
configuration to the client.

The algorithm is the classic exponential backoff algorithm. Everytime
we retry, we double the amount of time we might potentially wait. The
exact amount is randomly chosen within that range. The "base" amount
of wait (the amount for the first retry) is given by
configuration. Assuming every client participates in this scheme, it
should mean we'll back off and eventually match the backends max
capacity.

Reviewed By: edwinsmith

Differential Revision: D37448592

fbshipit-source-id: 3c20707b1b4527a4244a03db3d05453513bc113a