Decl_inherit: fix case where type constant inheritance does not match class constants
Summary: I caused a regression with {D36219663 (https://github.com/facebook/hhvm/commit/38a11e61a56d0b0b15c017db7a2d165ff142dfc2)} by tripping an older bug in `Decl_inherit`. For class constants, we would avoid replacing a non-synthesized inherited constant with a synthesized one, but we were missing this case for type constants. This is why in the added test case, the class constant override was raising an error, while the type constant was not, since the latter was filtered out of the map in Typing_extends.

Differential Revision: D37464394

fbshipit-source-id: a47fa331273a2b37a084613d72862e3902670929