Delete shallow class decls after we compute folded decls
Summary:
When the use_direct_decl_parser flag is on, parsing produces asts as well as
shallow decls into the shared heap.
Typechecking only uses folded class decls, so after we compute our
folded decls it should be safe to delete all the shallow class decls.

The hope is that this change will save enough memory for us to migrate
to the direct decl parser permanently, which will unblock Hack's deletion
of the old parser and also allow us to experiment further with
things unblocked by moving to the direct decl parser.

Reviewed By: fahndrich

Differential Revision: D37431324

fbshipit-source-id: c311f84c8129bf7d79d3fb1afeaacae0b87ef7a3