Improve expected types for `Hole`s in `ValCollection`
Summary:
When we have a `nullable` type in a `ValCollection` literal, we prefer to suggest the more specific inner type rather than `arraykey` if that inner type is a subtype of arraykey.

This diff ensures that we use this more specific type when available and use the supertype of all values in the collection otherwise.

FIXMEs in D37490630

Reviewed By: Wilfred

Differential Revision: D37488102

fbshipit-source-id: a6e5ab139f8a501b415a318b61e9fc1a6672a7a0