add missing span field to class_const node
Summary:
This field is missing and is inconsistent with other types of declarations. This is needed by Glean to properly index "class const".

I proceeded by imitating what was done for similar nodes, not 100% sure of the changes in `lowerer.rs` so reviewer should be extra vigilant there.

hack-dune and hhvm-and-hack-ubuntu are failing but this is unrelated to this diff

Reviewed By: Wilfred

Differential Revision: D37112943

fbshipit-source-id: 4a494ba317ac374ec84b08b5d94dd0d6954cc117