Fixed user_attr_args
Summary:
The user attributes printed in a function definition in HHAS are just a vec of typed values. Now that we have `assemble_typed_value`, can properly assemble user attributes.
Also added parsing function flags.
Also: A possible discrepancy is that there exists a `MEMOIZE_IMPL` `HhasFunctionFlag`, but it's not printed out by the `bytecode_printer`. So it may be dead code.

Reviewed By: hgoldstein

Differential Revision: D37466420

fbshipit-source-id: cfce0f93a6d32f2ae735b9f0014993ad5f599424