Intersect pushed like type with upper bound
Summary: The like-pushing subtyping rule requires that the types be well-formed, which prevents subtypes such as `dict<~string,bool> <: ~dict<string,bool>` because `~string` is not a subtype of the `arraykey` constraint on the first type parameter. By intersecting `~string` with `arraykey` we can ensure well-formedness without sacrificing error detection, as would happen if we used `dict<arraykey,bool> <: ~dict<string,bool>`.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D37426879

fbshipit-source-id: 2bb277a653b872df272e7f2c0d7c1b5b4cdc1380