Simplify Has_static_key constraints
Summary:
This diff changes the representation static key constraints. Instead of storing a number of fields, it now stores a single field with its type and its solely the solver's job to aggregate fields. This simplifies some of the treatment in the solver.

The diff breaks provenance because it was relying on empty shape maps being generated. There are other problems with provenance that needs redoing. The next diff will remove that feature for now.

Reviewed By: enetsee

Differential Revision: D37377639

fbshipit-source-id: 4e10a0f5465041b2b7c9f0daa06177a5aa828bf6