Infer optional fields
Summary:
The solver previously didn't account for optional fields (a field is used in one branch of control flow but not in the other). This diff introduces optional fields.

The approach is to generate optional field constraints at the point of explicit joins. We find all fields that flow into the incoming nodes of the join point and see which of the fields occur in both branches. If not, the field is optional.

Unlike other constraints we have, optional field constraints are not represented in the `constraint_` data type as they do not need to be generated while walking the program. They are internal to the solver.

Reviewed By: enetsee

Differential Revision: D37377636

fbshipit-source-id: 927dca0c6eeea7375dc591a3906f663fd2d203c7